requirements.txt
# FluxGym Requirements - Latest Working Versions (Sept 2025)
# Fixed compatibility issues and updated to latest stable packages

# Core ML and AI packages  
torch>=2.8.0
torchvision>=0.23.0
torchaudio>=2.8.0
transformers>=4.56.0
tokenizers>=0.22.0
safetensors>=0.4.5
huggingface-hub>=0.35.0
accelerate>=1.6.0
diffusers>=0.32.0

# Web Interface (Fixed versions for Gradio compatibility)
gradio
gradio-client
fastapi>=0.116.0
uvicorn>=0.36.0
# Note: gradio-logsview removed due to incompatibility - using standard Textbox

# Image and Data Processing
pillow>=10.4.0
numpy>=1.26.0,<2.0.0
pandas>=2.3.0
matplotlib>=3.10.0
opencv-python>=4.10.0
scipy>=1.11.0

# Configuration and Text Processing
pyyaml>=6.0.0  
toml>=0.10.2
python-slugify>=8.0.0
requests>=2.32.0
sentencepiece>=0.2.1
ftfy>=6.3.1

# Training and Optimization
bitsandbytes>=0.42.0
peft>=0.6.0
einops>=0.7.0
lion-pytorch>=0.2.3
prodigyopt>=1.1.2
lycoris-lora>=1.8.3

# Monitoring and Utils
tensorboard>=2.13.0
rich>=14.1.0
python-dotenv>=1.0.0
omegaconf>=2.3.0
voluptuous>=0.15.2

# Computer Vision
timm>=0.9.0
open_clip_torch>=2.20.0
k-diffusion>=0.0.15
albumentations>=1.3.0
kornia>=0.7.0
controlnet_aux>=0.0.7

# Quality metrics and extras
lpips>=0.1.4
pytorch_fid>=0.3.0
optimum-quanto>=0.1.0
invisible-watermark>=0.2.0
flatten_json>=0.1.0
imagesize>=1.4.1

# Version constraints
pydantic>=2.9.0,<3.0.0
hf_transfer>=0.1.0

# Windows compatibility
--prefer-binary
